Move classes into correct packages
